+++ dandelion/__init__.py
@@ -5,6 +5,7 @@ from . import update
 from . import activation
 from . import objective
 from . import functional
+from . import model
 
-__version__ = "0.14.4"
+__version__ = "0.15.1"
 __author__  = "David Leon (Dawei Leng)"